# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectorKind": {
        "enum": [
          "DRIVE",
          "GMAIL"
        ],
        "title": "ConnectorKind",
        "type": "string"
      },
...
      "ConnectorReauthorizationRead": {
        "properties": {
          "authorization_url": {
            "title": "Authorization Url",
            "type": "string"
          },
          "connector_id": {
            "format": "uuid",
            "title": "Connector Id",
            "type": "string"
          },
          "requested_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Requested Scopes",
            "type": "array"
          }
        },
        "required": [
          "connector_id",
          "authorization_url",
          "requested_scopes"
        ],
        "title": "ConnectorReauthorizationRead",
        "type": "object"
      },
      "ConnectorStatus": {
        "enum": [
          "ACTIVE",
          "REAUTH_REQUIRED",
          "ERROR"
        ],
        "title": "ConnectorStatus",
        "type": "string"
      },
      "ConnectorStatusRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ConnectorKind"
          },
          "requested_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Requested Scopes",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectorStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "updated_at",
          "requested_scopes"
        ],
        "title": "ConnectorStatusRead",
        "type": "object"
      },
...
      "EmailSummaryRead": {
        "properties": {
          "delivery_unknown": {
            "title": "Delivery Unknown",
            "type": "integer"
          },
          "retry_wait": {
            "title": "Retry Wait",
            "type": "integer"
          }
        },
        "required": [
          "retry_wait",
          "delivery_unknown"
        ],
        "title": "EmailSummaryRead",
        "type": "object"
      },
...
      "FailedJobRead": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/JobAction"
          },
          "action_resource_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Resource Id"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "kind",
          "state",
          "attempts",
          "error_code",
          "updated_at",
          "action",
          "action_resource_id"
        ],
        "title": "FailedJobRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAction": {
        "enum": [
          "RETRY_DRIVE_SYNC",
          "RETRY_EMAIL_DELIVERY",
          "RECONCILE_GMAIL",
          "NONE"
        ],
        "title": "JobAction",
        "type": "string"
      },
      "JobRetryRead": {
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "state",
          "version"
        ],
        "title": "JobRetryRead",
        "type": "object"
      },
      "JobSummaryRead": {
        "properties": {
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "queue_depth": {
            "title": "Queue Depth",
            "type": "integer"
          }
        },
        "required": [
          "queue_depth",
          "failed"
        ],
        "title": "JobSummaryRead",
        "type": "object"
      },
      "KnowledgeSourceStatusRead": {
        "properties": {
          "backlog": {
            "title": "Backlog",
            "type": "integer"
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "descendant_count": {
            "title": "Descendant Count",
            "type": "integer"
          },
          "include_descendants": {
            "title": "Include Descendants",
            "type": "boolean"
          },
          "isolated_files": {
            "title": "Isolated Files",
            "type": "integer"
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "recent_error_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recent Error Codes",
            "type": "array"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "root_folder_id": {
            "title": "Root Folder Id",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "status",
          "root_folder_id",
          "include_descendants",
          "descendant_count",
          "cursor",
          "last_success_at",
          "backlog",
          "isolated_files",
          "retry_count",
          "recent_error_codes"
        ],
        "title": "KnowledgeSourceStatusRead",
        "type": "object"
      },
      "OperationsSummaryRead": {
        "properties": {
          "connectors": {
            "items": {
              "$ref": "#/components/schemas/ConnectorStatusRead"
            },
            "title": "Connectors",
            "type": "array"
          },
          "email": {
            "$ref": "#/components/schemas/EmailSummaryRead"
          },
          "email_quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityStatusRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "jobs": {
            "$ref": "#/components/schemas/JobSummaryRead"
          },
          "knowledge_sources": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeSourceStatusRead"
            },
            "title": "Knowledge Sources",
            "type": "array"
          },
          "rag_quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityStatusRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "support": {
            "$ref": "#/components/schemas/SupportSummaryRead"
          }
        },
        "required": [
          "generated_at",
          "connectors",
          "knowledge_sources",
          "jobs",
          "support",
          "email",
          "rag_quality",
          "email_quality"
        ],
        "title": "OperationsSummaryRead",
        "type": "object"
      },
...
      "QualityStatusRead": {
        "properties": {
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "estimated_cost": {
            "title": "Estimated Cost",
            "type": "number"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "quality_score": {
            "title": "Quality Score",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "completed_at",
          "status",
          "quality_score",
          "latency_ms",
          "estimated_cost"
        ],
        "title": "QualityStatusRead",
        "type": "object"
      },
...
      "SupportSummaryRead": {
        "properties": {
          "backlog": {
            "title": "Backlog",
            "type": "integer"
          }
        },
        "required": [
          "backlog"
        ],
        "title": "SupportSummaryRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "PublicChatCredential": {
        "bearerFormat": "session credential",
        "description": "Opaque credential issued for one public chat session.",
        "scheme": "bearer",
        "type": "http"
      },
      "StaffSession": {
        "description": "Issued after approved staff OIDC admission; state-changing routes also require CSRF.",
        "in": "cookie",
        "name": "staff_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Enterprise AI Agent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/connectors/{connector_id}/reauthorize": {
      "post": {
        "operationId": "reauthorize_connector_api_v1_admin_connectors__connector_id__reauthorize_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connector Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorReauthorizationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Reauthorize Connector",
        "tags": [
          "administrator-operations"
        ],
        "x-security-classification": "administrator"
      }
    },
...
    "/api/v1/admin/jobs/failed": {
      "get": {
        "operationId": "failed_jobs_api_v1_admin_jobs_failed_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FailedJobRead"
                  },
                  "title": "Response Failed Jobs Api V1 Admin Jobs Failed Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Failed Jobs",
        "tags": [
          "administrator-operations"
        ],
        "x-security-classification": "administrator"
      }
    },
    "/api/v1/admin/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_v1_admin_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRetryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "administrator-operations"
        ],
        "x-security-classification": "administrator"
      }
    },
...
    "/api/v1/admin/operations/summary": {
      "get": {
        "operationId": "operations_summary_api_v1_admin_operations_summary_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsSummaryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Operations Summary",
        "tags": [
          "administrator-operations"
        ],
        "x-security-classification": "administrator"
      }
    },
...
  }
}